pyproject.toml
[tool.poetry]
name = "create_header-api"
version = "0.1.0"
description = ""
authors = ["Comamoca <comamoca.dev@gmail.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.79.0"
Pillow = "^9.2.0"
tweepy = "^4.10.0"
python-dotenv = "^0.20.0"
emoji = "1.7"
pilmoji = "^2.0.1"
loguru = "^0.6.0"
uvicorn = "^0.18.2"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
